Change sump light to turn off when main light comes on
LFS suggested that pH variance can be softened by having the sump light
running in a mostly opposite cycle than the main light should help
extend the respiration period.  Pulling more CO2 out of the water while
the coral in the tank are not consuming it.

Also adjusted AP700 about 4 inches to give a broader spread (similar to LFS).